Move Icon component to lib/
Re-usable components should go in here
