+++ oceannavigator/frontend/src/components/ComboBox.jsx
@@ -4,7 +4,7 @@ import React from "react";
 import $ from "jquery";
 import jQuery from "jquery";
 import {Modal, Button, FormControl} from "react-bootstrap";
-import Icon from "./Icon.jsx";
+import Icon from "./lib/Icon.jsx";
 import PropTypes from "prop-types";
 
 const i18n = require("../i18n.js");
