+++ oceannavigator/frontend/src/components/MapInputs.jsx
@@ -10,7 +10,7 @@ import Range from "./Range.jsx";
 import SelectBox from "./SelectBox.jsx";
 import DatasetSelector from "./DatasetSelector.jsx";
 import {Panel, Button, Row, Col, Tabs, Tab} from "react-bootstrap";
-import Icon from "./Icon.jsx";
+import Icon from "./lib/Icon.jsx";
 import Options from "./Options.jsx";
 import PropTypes from "prop-types";
