+++ oceannavigator/frontend/src/components/Permalink.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {Button, Well, Panel, Checkbox, Row} from "react-bootstrap";
-import Icon from "./Icon.jsx";
+import Icon from "./lib/Icon.jsx";
 import PropTypes from "prop-types";
 
 const i18n = require("../i18n.js");
